updated the code

+++ Tute01.c
@@ -5,17 +5,17 @@
 #include <stdio.h>
 
 int main() {
-  float mark1,mark2,avg;
-   
-   printf("\nEnter the marks:");
-   printf("\nMark 01:");
-   scanf("%f",mark1);
-   printf("\nMark 02:");
-   scanf("%f",mark2);
-   
-   avg=(float) (mark1+mark2)/2;
-   printf("\nAverage=%.1f",avg);
-   
+  double mark1,mark2,avg;
+
+printf("\nEnter the marks:");
+printf("\nMark 01:");
+scanf("%lf",mark1);
+printf("\nMark 02:");
+scanf("%lf",mark2);
+
+avg=(double)(mark1+mark2)/2;
+
+printf("\nAverage=%.1lf",avg);
   return 0;
 }
